+++ src/backends/postgres/migrations/2023-05-05-185144_deprecate_outliers/down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE IF EXISTS model
+    DROP COLUMN classification_id;